+++ ReactApp/client/run.js
@@ -1,25 +1,15 @@
 import thunk from 'redux-thunk'
 import { browserHistory } from 'react-router'
-import { routerMiddleware, routerReducer } from 'react-router-redux'
+import { routerMiddleware } from 'react-router-redux'
 
 //
 
 import { ReactApp } from '../index'
-import {
-    reducer as realtimeLocationReducer,
-    REALTIME_LOCATION_REDUCER_NAME,
-    actionUpdate
-} from '../../React/realtime-location'
-import {
-    reducerLocaleId as i18nReducerLocaleId,
-    reducerLocales as i18nReducerLocales,
-} from '../../i18n/redux'
+import { actionUpdate } from '../../React/realtime-location'
 import i18nRegister from '../../i18n/register/isomorphic.client'
+import { reducers } from '../../React/redux'
 
 //
-
-import { SERVER_REDUCER_NAME, serverReducer } from '../server/redux'
-const ROUTER_REDUCDER_NAME = 'routing'
 
 let logCountRouterUpdate = 0
 let logCountHistoryUpdate = 0
@@ -31,7 +21,7 @@ export default ({
     i18n = JSON.parse(process.env.KOOT_I18N) || false,
     router,
     redux,
-    store,
+    // store,
     client
 }) => {
     // ============================================================================
@@ -40,10 +30,12 @@ export default ({
 
     const reactApp = new ReactApp({ rootDom: 'root' })
 
-    reactApp.redux.middleware.use(thunk)
-    reactApp.redux.middleware.use(routerMiddleware(browserHistory))
-    // const routerHistory = browserHistory
-    // if (__CLIENT__) self.routerHistory = browserHistory
+    if (typeof redux.store !== 'undefined') {
+        reactApp.redux.middleware.use(thunk)
+        reactApp.redux.middleware.use(routerMiddleware(browserHistory))
+        // const routerHistory = browserHistory
+        // if (__CLIENT__) self.routerHistory = browserHistory
+    }
 
 
 
@@ -52,24 +44,11 @@ export default ({
     // Redux/Reducer 初始化
     // ============================================================================
 
-    const reducers = {
-        // 路由状态扩展
-        [ROUTER_REDUCDER_NAME]: routerReducer,
-        // 目的：新页面请求处理完成后再改变URL
-        [REALTIME_LOCATION_REDUCER_NAME]: realtimeLocationReducer,
-        // 对应服务器生成的store
-        [SERVER_REDUCER_NAME]: serverReducer,
-    }
-    if (i18n) {
-        reducers.localeId = i18nReducerLocaleId
-        reducers.locales = i18nReducerLocales
-    }
-
     // 兼容配置嵌套
     if (!redux)
         redux = client.redux
 
-    if (typeof store === 'undefined') {
+    if (typeof redux.store === 'undefined') {
         const { combineReducers } = redux
         if (typeof combineReducers === 'object') {
             for (let key in combineReducers) {
@@ -81,7 +60,7 @@ export default ({
             reactApp.redux.reducer.use(key, reducers[key])
         }
     } else {
-        reactApp.store = store
+        reactApp.store = redux.store
     }