+++ ReactApp/server/run.js
@@ -60,6 +60,7 @@ export default async (app, {
     locales = getLocalesDefault(),
     router,
     redux,
+    // store,
     client,
     server,
 }) => {
@@ -121,6 +122,7 @@ export default async (app, {
         i18n,
         router,
         redux,
+        // store,
         client
     })
     // if (__DEV__) console.log('├─ client code inited')
@@ -194,7 +196,8 @@ export default async (app, {
         routes: reactApp.react.router.get(),
 
         // redux store 对象
-        configStore: reactApp.createConfigureStoreFactory(),
+        configStore: typeof redux.store === 'undefined' ? reactApp.createConfigureStoreFactory() : undefined,
+        store: typeof redux.store === 'undefined' ? undefined : redux.store,
 
         // HTML基础模板
         template,