+++ index.js
@@ -1,9 +1,28 @@
+// Redux 相关
+
 /**
  * Redux Store
  * @type {Object}
  */
 export let store
 export const setStore = o => store = o
+
+/**
+ * 手动创建 Redux Store 时需要的相关对象
+ * @type {Object}
+ */
+import * as reduxForCreateStore from './React/redux'
+export { reduxForCreateStore }
+
+
+
+
+
+
+
+
+
+// 其他
 
 /**
  * History 对象